데이터베이스 조작끝

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -1,19 +1,13 @@
 package com.group.libraryapp.controller.user;
 
-import com.group.libraryapp.domain.user.User;
 import com.group.libraryapp.dto.user.UserCreateRequest;
 import com.group.libraryapp.dto.user.request.UserUpdateRequest;
 import com.group.libraryapp.dto.user.response.UserResponse;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowCallbackHandler;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.web.bind.annotation.*;
-import org.w3c.dom.stylesheets.LinkStyle;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
+
 
 @RestController
 public class UserController {
@@ -41,13 +35,25 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request){
+    String readSql="select * from user where id=?";
+    boolean isUserNotExist= jdbcTemplate.query(readSql,(rs ,rowNum)-> 0,
+        request.getId()).isEmpty();
+    if(isUserNotExist){
+        throw new IllegalArgumentException();
+    }
     String sql="update user set name= ? where id= ?";
     jdbcTemplate.update(sql,request.getName(),request.getId());
     }
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name){
+        String readSql="select * from user where name=?";
+        boolean isUserNotExist= jdbcTemplate.query(readSql, (rs,rowNum)-> 0, name).isEmpty();
+        if(isUserNotExist){
+            throw new IllegalArgumentException();
+        }
         String sql= "Delete from user Where name=?";
         jdbcTemplate.update(sql,name);
     }
+
 }
